Move `testing` submodule into a separate file
